+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # 1.0.0 (unreleased)
 
-Nothing yet. I don’t plan for there to be any real changes from 0.12.1;
+- Remove `bench` Cargo feature (by shifting benchmarks out of `src/lib.rs` into
+  `benches/bench.rs`; it still won’t run on anything but nightly, but that
+  don’t signify). Technically a [breaking-change], but it was something for
+  development only, so I’m not in the slightest bit concerned by it.
+
+I don’t plan for there to be any real changes from 0.12.1;
 it should be just a bit of housecleaning and a version bump.
 
 # 0.12.1 (2017-01-20)
