+++ Cargo.toml
@@ -9,6 +9,3 @@ repository = "https://github.com/chris-morgan/anymap"
 readme = "README.md"
 keywords = ["container", "data-structure", "map"]
 license = "MIT/Apache-2.0"
-
-[features]
-bench = []
